+++ miniproject/src/main/java/com/sparta/miniproject/service/PostService.java
@@ -3,7 +3,11 @@ package com.sparta.miniproject.service;
 import com.sparta.miniproject.S3.S3Service;
 import com.sparta.miniproject.dto.PostRequestDto;
 import com.sparta.miniproject.dto.PostResponseDto;
+import com.sparta.miniproject.exception.BusinessException;
+import com.sparta.miniproject.exception.ErrorCode;
 import com.sparta.miniproject.model.Post;
+import com.sparta.miniproject.model.User;
+import com.sparta.miniproject.repository.CommentRepository;
 import com.sparta.miniproject.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,6 +23,8 @@ import java.util.Objects;
 public class PostService {
 
     private final PostRepository postRepository;
+    private final CommentRepository commentRepository;
+    private final UserService userService;
     private final S3Service s3Service;
 
     // 포스트 리스트 조회
@@ -26,8 +32,10 @@ public class PostService {
         List<Post> posts = postRepository.findAllByOrderByCreatedAtDesc();
         List<PostResponseDto> postList = new ArrayList<>();
         for (Post post : posts) {
+            int countComment = commentRepository.countByPostId(post.getPostId());
             PostResponseDto postResponseDto = PostResponseDto.builder()
                     .post(post)
+                    .countComment(countComment)
                     .build();
             postList.add(postResponseDto);
         }
@@ -38,48 +46,69 @@ public class PostService {
     public PostResponseDto getPost(Long postId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("포스트가 존재하지 않습니다."));
+        int countComment = commentRepository.countByPostId(post.getPostId());
         return PostResponseDto.builder()
                 .post(post)
+                .countComment(countComment)
                 .build();
+    }
+
+    private String getUser() {
+        return userService.getMyInfo().getUsername();
     }
 
     // 포스트 생성
     @Transactional
     public Post createPost(PostRequestDto requestDto, MultipartFile imageFile) {
+        requestDto.setAuthor(getUser());
         String imagePath;
         if (!Objects.isNull(imageFile)) {
             try {
                 imagePath = s3Service.uploadImage(imageFile);
                 requestDto.setImgUrl(imagePath);
+                Post post = requestDto.createPost();
+                return postRepository.save(post);
             } catch (NullPointerException e) {
                 throw new NullPointerException("파일 변환에 실패했습니다");
             } catch (IllegalArgumentException e) {
                 throw new IllegalArgumentException(e.getMessage());
+            } catch (Exception e) {
+                Post post = requestDto.createPost();
+                return postRepository.save(post);
             }
         }
-        Post post = requestDto.createPost();
-        return postRepository.save(post);
+        return null;
     }
 
     //  포스트 수정
     @Transactional
-    public void updatePost(Long postId, PostRequestDto requestDto, MultipartFile file) {
+    public Post updatePost(Long postId, PostRequestDto requestDto, MultipartFile imageFile) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("포스트가 존재하지 않습니다."));
+                .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND));
+        if(!getUser().equals(post.getAuthor())) {
+            throw new BusinessException(ErrorCode.POST_UNAUTHORIZED);
+        }
         if(requestDto.getContents() == null || requestDto.getContents().equals("")){
-            throw new NullPointerException("제목과 내용을 채워주세요.");
+            throw new BusinessException(ErrorCode.LENGTH_REQUIRED);
         }
         String imagePath;
-        if (!Objects.isNull(file)) {
-            imagePath = s3Service.uploadImage(file);
+        if (!Objects.isNull(imageFile)) {
+            imagePath = s3Service.uploadImage(imageFile);
             requestDto.setImgUrl(imagePath);
         }
         post.update(requestDto.getContents(), requestDto.getImgUrl());
+        return post;
     }
 
-    // 포스트 삭제
-    public Long deletePost(Long postId) {
+        // 포스트 삭제
+        public Long deletePost(Long postId) throws IllegalArgumentException {
+            Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND));
+        if(!getUser().equals(post.getAuthor())) {
+            throw new BusinessException(ErrorCode.POST_UNAUTHORIZED);
+        }
         postRepository.deleteById(postId);
+        commentRepository.deleteByPostId(postId);
         return postId;
     }
 }
